# docs/openapi.json
{
  "swagger": "2.0",
  "basePath": "/",
  "paths": {
...
    "/v1/alunos/me": {
      "put": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AlunoResponse"
            }
          }
        },
        "summary": "Aluno atualiza seu perfil (Dados Pessoais + Endereço)",
        "operationId": "aluno_profile",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AlunoMeUpdateRequest"
            }
          }
        ],
        "tags": [
          "alunos"
        ]
      },
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Aluno exclui sua conta",
        "operationId": "aluno_delete",
        "tags": [
          "alunos"
        ]
      }
    },
...
    "/v1/users": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserListResponse"
            }
          },
          "403": {
            "description": "Forbidden - not a gestor"
          }
        },
        "summary": "Lista todos os usuários da prefeitura (Apenas Gestor)",
        "operationId": "list_users",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/alunos": {
      "post": {
        "responses": {
          "201": {
            "description": "Aluno account created"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Forbidden - not a gestor"
          },
          "409": {
            "description": "Conflict - duplicate email/CPF"
          }
        },
        "summary": "Gestor provisiona uma nova conta de aluno",
        "operationId": "aluno_provision_account",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AlunoProvisionAccountRequest"
            }
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/change-password": {
      "post": {
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "schema": {
              "$ref": "#/definitions/ChangePasswordResponse"
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Incorrect current password"
          },
          "429": {
            "description": "Too many attempts - rate limited"
          }
        },
        "summary": "Altera a senha do usuário logado (Requer senha atual)",
        "operationId": "change_user_password",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ChangePasswordRequest"
            }
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/fcm-token": {
      "patch": {
        "responses": {
          "200": {
            "description": "Token atualizado"
          },
          "400": {
            "description": "Token não enviado"
          }
        },
        "summary": "Atualiza o Token do Firebase (Push Notifications) do aparelho do usuário",
        "operationId": "update_fcm_token",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FcmTokenRequest"
            }
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/me": {
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Perfil do usuário logado",
        "operationId": "get_my_profile",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/motoristas": {
      "post": {
        "responses": {
          "201": {
            "description": "Motorista created",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Forbidden - not a gestor"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Gestor cria um novo Motorista",
        "operationId": "create_motorista",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/MotoristaCreateRequest"
            }
          }
        ],
        "tags": [
          "users"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserListResponse"
              }
            }
          },
          "403": {
            "description": "Forbidden - not a gestor"
          }
        },
        "summary": "Lista motoristas do mesmo município do Gestor",
        "operationId": "list_motoristas",
        "parameters": [
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/{id}": {
      "parameters": [
        {
          "in": "path",
          "description": "O UUID do usuário",
          "name": "id",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Busca usuário por ID (próprio perfil ou mesma prefeitura se Gestor)",
        "operationId": "get_user_by_id",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "BusKá API",
    "version": "1.0.0",
    "description": "\n## Sistema de Gerenciamento de Transporte Escolar\n\nAPI para gerenciamento de rotas, viagens e alunos do transporte escolar municipal.\n\n### Autenticação\nTodos os endpoints (exceto `/auth/login`) requerem autenticação JWT.\nInclua o header: `Authorization: Bearer <seu_token>`\n\n### Roles\n- **ALUNO**: Visualiza rotas, confirma presença em viagens\n- **MOTORISTA**: Gerencia viagens atribuídas, inicia/finaliza trajetos\n- **GESTOR**: Acesso completo à prefeitura (CRUD de rotas, motoristas, relatórios)\n        "
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT token. Format: Bearer <token>"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "definitions": {
...
    "UserListResponse": {
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserResponse"
          }
        },
        "total": {
          "type": "integer",
          "description": "Total de usuários"
        }
      },
      "type": "object"
    },
    "UserResponse": {
      "properties": {
        "id": {
          "type": "string",
          "description": "UUID do usuário"
        },
        "prefeitura_id": {
          "type": "string",
          "description": "UUID da prefeitura"
        },
        "nome": {
          "type": "string",
          "description": "Nome completo"
        },
        "email": {
          "type": "string",
          "description": "Email"
        },
        "telefone": {
          "type": "string",
          "description": "Telefone"
        },
        "cpf": {
          "type": "string",
          "description": "CPF"
        },
        "role": {
          "type": "string",
          "description": "Perfil (ALUNO, MOTORISTA, GESTOR)"
        },
        "status": {
          "type": "string",
          "description": "Status (PENDING_SIGNUP, ACTIVE, DISABLED)"
        },
        "signup_completed_at": {
          "type": "string",
          "format": "date-time",
          "description": "Data de conclusão do cadastro"
        },
        "municipio_nome": {
          "type": "string",
          "description": "Nome do município (prefeitura)"
        },
        "municipio_uf": {
          "type": "string",
          "description": "UF do município"
        },
        "matricula": {
          "type": "string",
          "description": "Matrícula (aluno)"
        },
        "nome_pai": {
          "type": "string",
          "description": "Nome do pai (aluno)"
        },
        "nome_mae": {
          "type": "string",
          "description": "Nome da mãe (aluno)"
        },
        "cnh": {
          "type": "string",
          "description": "CNH (motorista)"
        }
      },
      "type": "object"
    },
    "AlunoProvisionAccountRequest": {
      "required": [
        "cpf",
        "email",
        "nome",
        "password"
      ],
      "properties": {
        "nome": {
          "type": "string",
          "description": "Nome completo"
        },
        "email": {
          "type": "string",
          "description": "Email"
        },
        "password": {
          "type": "string",
          "description": "Senha"
        },
        "cpf": {
          "type": "string",
          "description": "CPF"
        },
        "telefone": {
          "type": "string",
          "description": "Telefone"
        }
      },
      "type": "object"
    },
    "MotoristaCreateRequest": {
      "required": [
        "cnh",
        "cpf",
        "email",
        "nome",
        "password"
      ],
      "properties": {
        "nome": {
          "type": "string",
          "description": "Nome completo"
        },
        "email": {
          "type": "string",
          "description": "Email"
        },
        "password": {
          "type": "string",
          "description": "Senha"
        },
        "cpf": {
          "type": "string",
          "description": "CPF"
        },
        "telefone": {
          "type": "string",
          "description": "Telefone"
        },
        "cnh": {
          "type": "string",
          "description": "CNH"
        },
        "salario": {
          "type": "number",
          "description": "Salário"
        }
      },
      "type": "object"
    },
    "ChangePasswordRequest": {
      "required": [
        "current_password",
        "new_password"
      ],
      "properties": {
        "current_password": {
          "type": "string",
          "description": "Senha atual"
        },
        "new_password": {
          "type": "string",
          "description": "Nova senha"
        }
      },
      "type": "object"
    },
    "ChangePasswordResponse": {
      "properties": {
        "message": {
          "type": "string",
          "description": "Mensagem de sucesso"
        }
      },
      "type": "object"
    },
    "FcmTokenRequest": {
      "required": [
        "fcm_token"
      ],
      "properties": {
        "fcm_token": {
          "type": "string",
          "description": "Token do dispositivo gerado pelo Firebase no Frontend"
        }
      },
      "type": "object"
    },
...
    "EnderecoInput": {
      "required": [
        "bairro",
        "cep",
        "cidade",
        "estado",
        "numero",
        "rua"
      ],
      "properties": {
        "rua": {
          "type": "string",
          "description": "Nome da rua"
        },
        "numero": {
          "type": "string",
          "description": "Número"
        },
        "bairro": {
          "type": "string",
          "description": "Bairro"
        },
        "cidade": {
          "type": "string",
          "description": "Cidade"
        },
        "estado": {
          "type": "string",
          "description": "Estado (UF)"
        },
        "cep": {
          "type": "string",
          "description": "CEP"
        },
        "complemento": {
          "type": "string",
          "description": "Complemento"
        }
      },
      "type": "object"
    },
    "AlunoResponse": {
      "properties": {
        "id": {
          "type": "string",
          "description": "UUID do aluno"
        },
        "nome": {
          "type": "string",
          "description": "Nome completo"
        },
        "matricula": {
          "type": "string",
          "description": "Matrícula"
        },
        "escola": {
          "type": "string",
          "description": "Nome da escola"
        },
        "status": {
          "type": "string",
          "description": "Status (PENDING_SIGNUP, ACTIVE, DISABLED)"
        },
        "signup_completed_at": {
          "type": "string",
          "format": "date-time",
          "description": "Data de conclusão do cadastro"
        }
      },
      "type": "object"
    },
    "AlunoMeUpdateRequest": {
      "properties": {
        "nome": {
          "type": "string",
          "description": "Nome completo"
        },
        "telefone": {
          "type": "string",
          "description": "Telefone"
        },
        "matricula": {
          "type": "string",
          "description": "Matrícula escolar"
        },
        "nome_pai": {
          "type": "string",
          "description": "Nome do pai"
        },
        "cpf_pai": {
          "type": "string",
          "description": "CPF do pai"
        },
        "nome_mae": {
          "type": "string",
          "description": "Nome da mãe"
        },
        "cpf_mae": {
          "type": "string",
          "description": "CPF da mãe"
        },
        "endereco_casa": {
          "$ref": "#/definitions/EnderecoInput"
        }
      },
      "type": "object"
    },
...
  },
...
}